Namespacing improvements and event/error signaling improvements
These changes would ideally be split in to lots of small commits, but it got
to a point where that didn't seem practical, so there are lots of misc
changes in this changeset.

The namespacing code in gx-gen got an overhaul to generalize certain
problems so they can be dealt with in one place, and improve the consistency
of our generated symbols.

This moves towards extensions being built into seperate libraries, each with
their own _init() function.

Quite a bit of work was done relating to event and error handling, including
generating various tables for events that allow us to do runtime lookups for
meta data; such as a description of an event, or lookup a window offset for
any events directly related to windows.

Signal handling of expose events now works - see test-hello-world.c